进一步完善界面UI
1. 动画效果结束时的回调函数，改为Promise的then方式；
2. UI需要翻译的文本，尽量迁移到C++中使用LOCTEXT实现；目前把等级相关的部分先迁移；
3. 新增prompt_view，用于显示需要重点提示的信息，比如获取到探索点；
4. 完善MainPlayerState中关于探索点和等级的逻辑，以及实现”已升级过的等级不再消耗探索点“的设定；
5. 在菜单主页的右上角新增显示探索点数；

+++ Source/escape_solar_system/Public/BodyComponent.h
@@ -28,7 +28,7 @@ public:
 	void SetShieldHeat(EPawnType Type, int32 Level);
 	void SetShieldPress(EPawnType Type, int32 Level);
 
-	void ChangeHP(float Delta);
+	void ChangeHP(int32 Delta);
 	FORCEINLINE float GetMass() const { return Mass; }
 	FORCEINLINE float GetCurrentHP() const { return CurrentHP; }
 	FORCEINLINE float GetMaximumHP() const { return MaximumHP; }
